Time: 4 ms (17.2%), Space: 62.3 MB (32.26%) - LeetHub

+++ 3152-maximum-value-of-an-ordered-triplet-ii/3152-maximum-value-of-an-ordered-triplet-ii.java
@@ -14,7 +14,7 @@ suff_max[n-1]=nums[n-1];
        }
 
        for(int j=1;j<nums.length-1;j++){
-            ans=Math.max(ans,(long)((long)(pref_max[j-1]-nums[j])*suff_max[j+1]));
+            ans=Math.max(ans,((long)(pref_max[j-1]-nums[j])*suff_max[j+1]));
        }
        return ans;
     }
